Rename ContactContract to AddressContract
