+++ core/src/androidTest/kotlin/at/bitfire/davdroid/resource/LocalGroupTest.kt
@@ -15,8 +15,8 @@ import android.provider.ContactsContract.CommonDataKinds.GroupMembership
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.rule.GrantPermissionRule
 import at.bitfire.synctools.mapping.contacts.Contact
-import at.bitfire.synctools.storage.contacts.ContactContract.CachedGroupMembership
-import at.bitfire.synctools.storage.contacts.ContactContract.asSyncAdapter
+import at.bitfire.synctools.storage.contacts.AddressContract.CachedGroupMembership
+import at.bitfire.synctools.storage.contacts.AddressContract.asSyncAdapter
 import at.bitfire.synctools.storage.contacts.ContactsBatchOperation
 import at.bitfire.synctools.vcard.GroupMethod
 import dagger.hilt.android.qualifiers.ApplicationContext
