+++ core/src/main/kotlin/at/bitfire/davdroid/ui/DebugInfoGenerator.kt
@@ -59,7 +59,7 @@ import java.util.logging.Logger
 import javax.inject.Inject
 import kotlin.use
 import at.bitfire.synctools.storage.calendar.EventsContract.asSyncAdapter as asCalendarSyncAdapter
-import at.bitfire.synctools.storage.contacts.ContactContract.asSyncAdapter as asContactsSyncAdapter
+import at.bitfire.synctools.storage.contacts.AddressContract.asSyncAdapter as asContactsSyncAdapter
 import at.techbee.jtx.JtxContract.asSyncAdapter as asJtxSyncAdapter
 
 @WorkerThread
