+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/contacts/builder/PhotoBuilder.kt
@@ -13,7 +13,7 @@ import android.provider.ContactsContract.CommonDataKinds.Photo
 import android.provider.ContactsContract.RawContacts
 import at.bitfire.synctools.mapping.contacts.Contact
 import at.bitfire.synctools.storage.BatchOperation
-import at.bitfire.synctools.storage.contacts.ContactContract.asSyncAdapter
+import at.bitfire.synctools.storage.contacts.AddressContract.asSyncAdapter
 import java.io.IOException
 import java.util.logging.Level
 import java.util.logging.Logger
