+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/contacts/builder/UnknownPropertiesBuilder.kt
@@ -7,7 +7,7 @@ package at.bitfire.synctools.mapping.contacts.builder
 import android.net.Uri
 import at.bitfire.synctools.mapping.contacts.Contact
 import at.bitfire.synctools.storage.BatchOperation
-import at.bitfire.synctools.storage.contacts.ContactContract
+import at.bitfire.synctools.storage.contacts.AddressContract
 import java.util.LinkedList
 
 class UnknownPropertiesBuilder(dataRowUri: Uri, rawContactId: Long?, contact: Contact, readOnly: Boolean) :
@@ -16,14 +16,14 @@ class UnknownPropertiesBuilder(dataRowUri: Uri, rawContactId: Long?, contact: Co
     override fun build(): List<BatchOperation.CpoBuilder> {
         val result = LinkedList<BatchOperation.CpoBuilder>()
         contact.unknownProperties?.let { unknownProperties ->
-            result += newDataRow().withValue(ContactContract.UnknownProperty.UNKNOWN_PROPERTIES, unknownProperties)
+            result += newDataRow().withValue(AddressContract.UnknownProperty.UNKNOWN_PROPERTIES, unknownProperties)
         }
         return result
     }
 
 
     object Factory : DataRowBuilder.Factory<UnknownPropertiesBuilder> {
-        override fun mimeType() = ContactContract.UnknownProperty.CONTENT_ITEM_TYPE
+        override fun mimeType() = AddressContract.UnknownProperty.CONTENT_ITEM_TYPE
         override fun newInstance(dataRowUri: Uri, rawContactId: Long?, contact: Contact, readOnly: Boolean) =
             UnknownPropertiesBuilder(dataRowUri, rawContactId, contact, readOnly)
     }
