+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/contacts/handler/UnknownPropertiesHandler.kt
@@ -6,16 +6,16 @@ package at.bitfire.synctools.mapping.contacts.handler
 
 import android.content.ContentValues
 import at.bitfire.synctools.mapping.contacts.Contact
-import at.bitfire.synctools.storage.contacts.ContactContract
+import at.bitfire.synctools.storage.contacts.AddressContract
 
 object UnknownPropertiesHandler : DataRowHandler() {
 
-    override fun forMimeType() = ContactContract.UnknownProperty.CONTENT_ITEM_TYPE
+    override fun forMimeType() = AddressContract.UnknownProperty.CONTENT_ITEM_TYPE
 
     override fun handle(values: ContentValues, contact: Contact) {
         super.handle(values, contact)
 
-        contact.unknownProperties = values.getAsString(ContactContract.UnknownProperty.UNKNOWN_PROPERTIES)
+        contact.unknownProperties = values.getAsString(AddressContract.UnknownProperty.UNKNOWN_PROPERTIES)
     }
 
 }
