+++ synctools/src/main/kotlin/at/bitfire/synctools/storage/contacts/AndroidAddressBook.kt
@@ -13,7 +13,7 @@ import android.provider.ContactsContract
 import android.provider.ContactsContract.Groups
 import android.provider.ContactsContract.RawContacts
 import androidx.core.content.contentValuesOf
-import at.bitfire.synctools.storage.contacts.ContactContract.asSyncAdapter
+import at.bitfire.synctools.storage.contacts.AddressContract.asSyncAdapter
 import at.bitfire.synctools.storage.toContentValues
 import at.bitfire.synctools.vcard.GroupMethod
 import java.io.FileNotFoundException
