+++ synctools/src/main/kotlin/at/bitfire/synctools/storage/contacts/AndroidGroup.kt
@@ -17,7 +17,7 @@ import androidx.annotation.CallSuper
 import androidx.core.content.contentValuesOf
 import at.bitfire.synctools.mapping.contacts.Contact
 import at.bitfire.synctools.storage.LocalStorageException
-import at.bitfire.synctools.storage.contacts.ContactContract.asSyncAdapter
+import at.bitfire.synctools.storage.contacts.AddressContract.asSyncAdapter
 import at.bitfire.synctools.storage.plusAssign
 import java.io.FileNotFoundException
 import java.util.logging.Logger
