+++ synctools/src/test/kotlin/at/bitfire/synctools/mapping/contacts/builder/UnknownPropertiesBuilderTest.kt
@@ -6,7 +6,7 @@ package at.bitfire.synctools.mapping.contacts.builder
 
 import android.net.Uri
 import at.bitfire.synctools.mapping.contacts.Contact
-import at.bitfire.synctools.storage.contacts.ContactContract
+import at.bitfire.synctools.storage.contacts.AddressContract
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -28,8 +28,8 @@ class UnknownPropertiesBuilderTest {
             unknownProperties = "X-TEST:12345"
         }, false).build().also { result ->
             assertEquals(1, result.size)
-            assertEquals(ContactContract.UnknownProperty.CONTENT_ITEM_TYPE, result[0].values[ContactContract.UnknownProperty.MIMETYPE])
-            assertEquals("X-TEST:12345", result[0].values[ContactContract.UnknownProperty.UNKNOWN_PROPERTIES])
+            assertEquals(AddressContract.UnknownProperty.CONTENT_ITEM_TYPE, result[0].values[AddressContract.UnknownProperty.MIMETYPE])
+            assertEquals("X-TEST:12345", result[0].values[AddressContract.UnknownProperty.UNKNOWN_PROPERTIES])
         }
     }
