+++ synctools/src/test/kotlin/at/bitfire/synctools/mapping/contacts/handler/UnknownPropertiesHandlerTest.kt
@@ -6,7 +6,7 @@ package at.bitfire.synctools.mapping.contacts.handler
 
 import android.content.ContentValues
 import at.bitfire.synctools.mapping.contacts.Contact
-import at.bitfire.synctools.storage.contacts.ContactContract
+import at.bitfire.synctools.storage.contacts.AddressContract
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
@@ -20,7 +20,7 @@ class UnknownPropertiesHandlerTest {
     fun testUnknownProperties_Empty() {
         val contact = Contact()
         UnknownPropertiesHandler.handle(ContentValues().apply {
-            putNull(ContactContract.UnknownProperty.UNKNOWN_PROPERTIES)
+            putNull(AddressContract.UnknownProperty.UNKNOWN_PROPERTIES)
         }, contact)
         assertNull(contact.unknownProperties)
     }
@@ -29,7 +29,7 @@ class UnknownPropertiesHandlerTest {
     fun testUnknownProperties_Values() {
         val contact = Contact()
         UnknownPropertiesHandler.handle(ContentValues().apply {
-            put(ContactContract.UnknownProperty.UNKNOWN_PROPERTIES, "X-TEST:12345")
+            put(AddressContract.UnknownProperty.UNKNOWN_PROPERTIES, "X-TEST:12345")
         }, contact)
         assertEquals("X-TEST:12345", contact.unknownProperties)
     }
